update existing map_item tests and add helper
